Moved concepts to types and util namespace
Bring a bit more order to this

[CURA-10724]
