+++ include/utils/types/graph.h
@@ -7,9 +7,9 @@
 #include <concepts>
 #include <type_traits>
 
-#include "utils/concepts/generic.h"
+#include "utils/types/generic.h"
 
-namespace cura
+namespace cura::utils
 {
 
 /* # nodable
